Update list components directory layout
